Move to obs_ops.splits
